delete one user when accepting bubble reco notification

+++ controllers/notificationController.js
@@ -73,6 +73,16 @@ export const acceptNotification = async (req, res) => {
       return res.status(200).json({ reco, message: "deleted notification" });
     }
     if (notification.type === "NOTIFICATION_ABOUT_RECO_IN_BUBBLE") {
+      const notification = await Notification.findById(notificationId);
+      if (notification.userIds.length > 1) {
+        await Notification.updateOne(
+          { _id: notificationId },
+          { $pull: { userIds: id } }
+        );
+        return res
+          .status(205)
+          .json({ message: "removed user from notification" });
+      }
       await Notification.findByIdAndDelete(notificationId);
       return res.status(205).json({ message: "deleted notification" });
     }
